Seleniumpython/Seleniumpython/tests/requirement.txt
attrs==24.2.0
beautifulsoup4==4.12.3
behave==1.2.6
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.15.4
h11==0.14.0
idna==3.7
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
outcome==1.3.0.post0
packaging==24.1
parameterized==0.9.0
parse==1.20.2
parse-type==0.6.2
pdbp==1.5.4
platformdirs==4.2.2
pluggy==1.5.0
py==1.11.0
Pygments==2.18.0
pynose==1.5.2
pyotp==2.9.0
PySocks==1.7.1
pytest==8.3.2
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
PyYAML==6.0.2
requests==2.31.0
rich==13.7.1
sbvirtualdisplay==1.3.0
selenium==4.23.1
seleniumbase==4.29.9
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tabcompleter==1.3.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
